# 🚀 Major UX Overhaul: Workflow-Centric Design & Authentication Fixes